refactor(gov): msg <-> resp
