+++ contracts/pylon/gov/src/testing/test_poll_snapshot.rs
@@ -1,6 +1,6 @@
 use cosmwasm_std::testing::{mock_env, mock_info, MOCK_CONTRACT_ADDR};
 use cosmwasm_std::{attr, coins, Uint128};
-use pylon_token::gov::{ExecuteMsg, PollMsg};
+use pylon_token::gov_msg::{ExecuteMsg, PollMsg};
 
 use crate::contract;
 use crate::error::ContractError;
