+++ packages/pylon_token/src/lib.rs
@@ -3,6 +3,7 @@ pub mod collector;
 pub mod common;
 pub mod community;
 pub mod distributor;
-pub mod gov;
+pub mod gov_msg;
+pub mod gov_resp;
 pub mod staking;
 pub mod vesting;
